move sidebar files to sidebar folder
