+++ src/modules/ui/layout/index.ts
@@ -1,4 +1,4 @@
-import Sidebar from "./Sidebar.vue";
+import Sidebar from "./sidebar/Sidebar.vue";
 import Topbar from "./Topbar.vue";
 
 export { Sidebar, Topbar };
