+++ src/modules/ui/layout/sidebar/Sidebar.vue
@@ -37,7 +37,7 @@ import { AUTH_MODULE_KEY, injectStrict } from "@/utils/inject";
 import SidebarLogo from "./SidebarLogo.vue";
 import SidebarItem from "./SidebarItem.vue";
 import SidebarCategoryItem from "./SidebarCategoryItem.vue";
-import { SidebarGroupItem, SidebarSingleItem, SidebarItems } from "./types";
+import { SidebarGroupItem, SidebarSingleItem, SidebarItems } from "../types";
 import { useSidebarItems } from "./SidebarItems.composable";
 
 const authModule = injectStrict(AUTH_MODULE_KEY);
