+++ src/modules/ui/layout/sidebar/SidebarCategoryItem.unit.ts
