+++ src/modules/ui/layout/sidebar/SidebarLogo.unit.ts
