+++ src/modules/ui/layout/sidebar/SidebarLogo.vue
